+++ src/Factorial/GUIFactorial.java
@@ -6,12 +6,12 @@ public class GUIFactorial {
 		return "Factorial"; 
 	}
 	
-	public static int iterative(int num) {
+	public static double iterative(int num) {
         if (num < 0) {
             throw new IllegalArgumentException("Factorial function is only available for numbers greater than 0.");
         }
 
-        int prod = 1;
+        double prod = 1;
         for (int i = 2; i <= num; i++) {
             prod = prod * i;
         }
@@ -19,7 +19,7 @@ public class GUIFactorial {
         return prod;
     }
 	
-	public static int recursive(int num) {
+	public static double recursive(int num) {
         if (num < 0) {
             throw new IllegalArgumentException("Factorial function is only available for numbers greater than 0.");
         }
@@ -33,7 +33,7 @@ public class GUIFactorial {
     * @param The number that we need to compute a factorial for
     * @return The factorial of the number
      */
-    private static int computeFactorial(int num) {
+    private static double computeFactorial(int num) {
        if (num <= 1) {
            return 1;
        }
